configure mongodbmemoryserver and jest for testing purposes
