+++ auth/package.json
@@ -4,7 +4,15 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "ts-node-dev src/index.ts"
+    "start": "ts-node-dev src/index.ts",
+    "test": "jest --watchAll --no-cache"
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "testEnvironment": "node",
+    "setupFilesAfterEnv": [
+      "./src/test/setup.ts"
+    ]
   },
   "keywords": [],
   "author": "Jimmy K Njuguna <jameskinyua590@gmail.com>",
@@ -23,5 +31,13 @@
     "mongoose": "^5.10.0",
     "ts-node-dev": "^1.0.0-pre.56",
     "typescript": "^3.9.7"
+  },
+  "devDependencies": {
+    "@types/jest": "^26.0.10",
+    "@types/supertest": "^2.0.10",
+    "jest": "^26.4.2",
+    "mongodb-memory-server": "^6.6.7",
+    "supertest": "^4.0.2",
+    "ts-jest": "^26.3.0"
   }
 }
