+++ auth/src/index.ts
@@ -1,39 +1,8 @@
-import express from "express";
-import "express-async-errors";
-import { json } from "body-parser";
 import mongoose from "mongoose";
-import cookieSession from "cookie-session";
 
-import { currentUserRouter } from "./routes/current-user";
-import { signinRouter } from "./routes/signin";
-import { signoutRouter } from "./routes/signout";
-import { signupRouter } from "./routes/signup";
-import { errorHandler } from "./middlewares/error-handler";
-import { NotFoundError } from "./errors/not-found-error";
+import { app } from "./app";
 
-const app = express();
 const port = 3000;
-
-app.set("trust proxy", true);
-
-app.use(json());
-app.use(
-  cookieSession({
-    signed: false,
-    secure: true,
-  })
-);
-
-app.use(currentUserRouter);
-app.use(signinRouter);
-app.use(signoutRouter);
-app.use(signupRouter);
-
-app.all("*", (req, res) => {
-  throw new NotFoundError();
-});
-
-app.use(errorHandler);
 
 const start = async () => {
   if (!process.env.JWT_KEY) {
